# frontend/src/lib/api/openapi.json
{"openapi":"3.1.0","info":{"title":"Work Order API","version":"0.1.0"},"paths":{"/api/v1/auth/signup":{"post":{"tags":["Authentication"],"summary":"Sign Up","description":"Register a new user.","operationId":"sign_up_api_v1_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignUp"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/signin":{"post":{"tags":["Authentication"],"summary":"Sign In","description":"Sign in an existing user.","operationId":"sign_in_api_v1_auth_signin_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/signout":{"post":{"tags":["Authentication"],"summary":"Sign Out","description":"Sign out the current user.","operationId":"sign_out_api_v1_auth_signout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Me","description":"Get current user information.","operationId":"get_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Me Api V1 Auth Me Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh Token","description":"Refresh access token.","operationId":"refresh_token_api_v1_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenRefresh"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/locations":{"get":{"tags":["locations"],"summary":"Get Locations","description":"Get all locations.","operationId":"get_locations_api_v1_locations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Location"},"type":"array","title":"Response Get Locations Api V1 Locations Get"}}}}},"security":[{"HTTPBearer":[]},{"HTTPBearer":[]}]},"post":{"tags":["locations"],"summary":"Create Location","description":"Create a new location.","operationId":"create_location_api_v1_locations_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]},{"HTTPBearer":[]}]}},"/api/v1/locations/{location_id}":{"get":{"tags":["locations"],"summary":"Get Location","description":"Get a specific location by ID.","operationId":"get_location_api_v1_locations__location_id__get","security":[{"HTTPBearer":[]},{"HTTPBearer":[]}],"parameters":[{"name":"location_id","in":"path","required":true,"schema":{"type":"string","title":"Location Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/work-orders":{"get":{"tags":["work-orders"],"summary":"Get Work Orders","description":"Get work orders with optional filtering and pagination.","operationId":"get_work_orders_api_v1_work_orders_get","security":[{"HTTPBearer":[]},{"HTTPBearer":[]}],"parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"priority","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Priority"}},{"name":"assigned_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Assigned To"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrdersResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["work-orders"],"summary":"Create Work Order","description":"Create a new work order.","operationId":"create_work_order_api_v1_work_orders_post","security":[{"HTTPBearer":[]},{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrderCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrder"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/work-orders/{work_order_id}":{"get":{"tags":["work-orders"],"summary":"Get Work Order","description":"Get a specific work order by ID.","operationId":"get_work_order_api_v1_work_orders__work_order_id__get","security":[{"HTTPBearer":[]},{"HTTPBearer":[]}],"parameters":[{"name":"work_order_id","in":"path","required":true,"schema":{"type":"string","title":"Work Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrder"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["work-orders"],"summary":"Update Work Order","description":"Update a work order.","operationId":"update_work_order_api_v1_work_orders__work_order_id__put","security":[{"HTTPBearer":[]},{"HTTPBearer":[]}],"parameters":[{"name":"work_order_id","in":"path","required":true,"schema":{"type":"string","title":"Work Order Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrderUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrder"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["work-orders"],"summary":"Delete Work Order","description":"Delete a work order.","operationId":"delete_work_order_api_v1_work_orders__work_order_id__delete","security":[{"HTTPBearer":[]},{"HTTPBearer":[]}],"parameters":[{"name":"work_order_id","in":"path","required":true,"schema":{"type":"string","title":"Work Order Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Root  Get"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}}},"components":{"schemas":{"AuthResponse":{"properties":{"user":{"$ref":"#/components/schemas/UserResponse"},"session":{"anyOf":[{"$ref":"#/components/schemas/SessionResponse"},{"type":"null"}]}},"type":"object","required":["user"],"title":"AuthResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Location":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state_province":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State Province"},"postal_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Postal Code"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"latitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latitude"},"longitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Longitude"}},"type":"object","required":["id","name"],"title":"Location"},"LocationCreate":{"properties":{"name":{"type":"string","title":"Name"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state_province":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State Province"},"postal_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Postal Code"},"country":{"type":"string","title":"Country","default":"USA"},"latitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latitude"},"longitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Longitude"}},"type":"object","required":["name"],"title":"LocationCreate"},"MessageResponse":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"MessageResponse"},"PaginationInfo":{"properties":{"page":{"type":"integer","title":"Page"},"limit":{"type":"integer","title":"Limit"},"total":{"type":"integer","title":"Total"},"totalPages":{"type":"integer","title":"Totalpages"}},"type":"object","required":["page","limit","total","totalPages"],"title":"PaginationInfo"},"SessionResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"expires_in":{"type":"integer","title":"Expires In"},"expires_at":{"type":"integer","title":"Expires At"}},"type":"object","required":["access_token","refresh_token","expires_in","expires_at"],"title":"SessionResponse"},"TokenRefresh":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"TokenRefresh"},"UserResponse":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","title":"Email"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Updated At"},"email_confirmed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email Confirmed At"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"last_sign_in_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Sign In At"}},"type":"object","required":["id","email","created_at"],"title":"UserResponse"},"UserSignIn":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserSignIn"},"UserSignUp":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserSignUp"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WorkOrder":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"status":{"$ref":"#/components/schemas/WorkOrderStatus"},"priority":{"$ref":"#/components/schemas/WorkOrderPriority"},"location":{"anyOf":[{"$ref":"#/components/schemas/Location"},{"type":"null"}]},"location_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location Id"},"assigned_to_user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Assigned To User Id"},"created_by_user_id":{"type":"string","title":"Created By User Id"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"}},"type":"object","required":["id","title","status","priority","created_by_user_id","created_at","updated_at"],"title":"WorkOrder"},"WorkOrderCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"status":{"$ref":"#/components/schemas/WorkOrderStatus","default":"Open"},"priority":{"$ref":"#/components/schemas/WorkOrderPriority","default":"Medium"},"location_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location Id"},"assigned_to_user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Assigned To User Id"}},"type":"object","required":["title"],"title":"WorkOrderCreate"},"WorkOrderPriority":{"type":"string","enum":["Low","Medium","High"],"title":"WorkOrderPriority"},"WorkOrderStatus":{"type":"string","enum":["Open","In Progress","Completed","Cancelled","On Hold"],"title":"WorkOrderStatus"},"WorkOrderUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"status":{"anyOf":[{"$ref":"#/components/schemas/WorkOrderStatus"},{"type":"null"}]},"priority":{"anyOf":[{"$ref":"#/components/schemas/WorkOrderPriority"},{"type":"null"}]},"location_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location Id"},"assigned_to_user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Assigned To User Id"}},"type":"object","title":"WorkOrderUpdate"},"WorkOrdersResponse":{"properties":{"data":{"items":{"$ref":"#/components/schemas/WorkOrder"},"type":"array","title":"Data"},"count":{"type":"integer","title":"Count"},"pagination":{"anyOf":[{"$ref":"#/components/schemas/PaginationInfo"},{"type":"null"}]}},"type":"object","required":["data","count"],"title":"WorkOrdersResponse"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}